requirements.txt
Django==3.1.4
django-phonenumber-field==5.0.0
django-redis==4.12.1
djangorestframework==3.12.2
phonenumbers==8.12.15
redis==3.5.3
requests==2.25.1
